pyproject.toml
[tool.poetry]
name = "calma"
version = "0.1.0"
description = ""
authors = ["WellyntonJoãoMartins <wellyntonjoaomartins@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.4"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.7"}


[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.0"

[tool.taskipy.tasks]
dev = "fastapi dev dev.py"
prod = "uvicorn src.api.application:app --host 0.0.0.0 --port 8000 --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
